[MOD] Adding a choice when getting the next floor to stay on the same

+++ src/Battle.c
@@ -25,7 +25,7 @@ void CloseBattleWindow(void)
 	PopMenu();
 }
 
-static int currentFloor = 1;
+static uint8_t currentFloor = 1;
 
 void ResetFloor(void)
 {
